requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.7.2
attrs==23.2.0
Babel==2.15.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
distlib==0.3.8
Django==5.0.1
django-autoslug==1.9.9
django-ipware==7.0.1
django-localflavor==4.0
django-money==3.5.2
django-money-rates==0.3.1
django-tinymce==3.7.1
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.1
idna==3.7
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
maxminddb==2.6.2
multidict==6.0.5
opencv-python==4.10.0.84
packaging==24.2
pefile==2023.2.7
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
psutil==6.0.0
ptyprocess==0.7.0
py-moneyed==3.0
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMySQL==1.1.0
pypiwin32==223
python-ipware==3.0.0
python-stdnum==1.19
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==71.1.0
sh==2.2.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
virtualenv==20.25.0
whitenoise==6.7.0
Whoosh==2.7.4
yarl==1.9.4
yt-dlp==2025.1.15
